Refactor: page-based file layout
Split src/components/ by route. Page-specific UI moves into
src/pages/<route>/ (atlas, about, privacy, roadmap, sources,
design-lab); only genuinely cross-page primitives — ui.tsx and
audit/ — stay in src/components/.

Atlas absorbs its full chrome (BudgetExplorer + Masthead, PageNav,
ShareLink, PitWarning, plus the 10 section components), since none
of them are imported by the secondary pages.

src/lib/ and src/data/ untouched. CLAUDE.md updated to document
the new layout.
